+++ src/account/default.ts
@@ -1,5 +1,3 @@
-import { BN } from 'bn.js';
-
 import { UDC, ZERO } from '../constants';
 import { ProviderInterface, ProviderOptions } from '../provider';
 import { Provider } from '../provider/default';
@@ -11,22 +9,27 @@ import {
   Call,
   DeclareContractPayload,
   DeclareContractResponse,
+  DeclareContractTransaction,
   DeclareDeployContractPayload,
   DeclareDeployUDCResponse,
   DeployAccountContractPayload,
+  DeployAccountContractTransaction,
   DeployContractResponse,
   DeployContractUDCResponse,
   EstimateFee,
   EstimateFeeAction,
   EstimateFeeDetails,
+  Invocation,
   InvocationsDetails,
   InvocationsSignerDetails,
   InvokeFunctionResponse,
   KeyPair,
   MultiDeployContractResponse,
   Signature,
+  TransactionBulk,
   UniversalDeployerContractPayload,
 } from '../types';
+import { EstimateFeeBulk, TransactionSimulation } from '../types/account';
 import { parseUDCEvent } from '../utils/events';
 import {
   calculateContractAddressFromHash,
@@ -34,10 +37,9 @@ import {
   pedersen,
   transactionVersion,
 } from '../utils/hash';
-import { BigNumberish, hexToDecimalString, toBN, toCairoBool } from '../utils/number';
+import { BigNumberish, toBN, toCairoBool } from '../utils/number';
 import { parseContract } from '../utils/provider';
 import { compileCalldata, estimatedFeeToMaxFee, randomAddress } from '../utils/stark';
-import { getStarknetIdContract, useDecoded, useEncoded } from '../utils/starknetId';
 import { fromCallsToExecuteCalldata } from '../utils/transaction';
 import { TypedData, getMessageHash } from '../utils/typedData';
 import { AccountInterface } from './interface';
@@ -60,52 +62,6 @@ export class Account extends Provider implements AccountInterface {
 
   public async getNonce(blockIdentifier?: BlockIdentifier): Promise<BigNumberish> {
     return super.getNonceForAddress(this.address, blockIdentifier);
-  }
-
-  public async getStarkName(StarknetIdContract?: string): Promise<string | Error> {
-    const chainId = await this.getChainId();
-    const contract = StarknetIdContract ?? getStarknetIdContract(chainId);
-
-    try {
-      const hexDomain = await this.callContract({
-        contractAddress: contract,
-        entrypoint: 'address_to_domain',
-        calldata: compileCalldata({
-          address: this.address,
-        }),
-      });
-      const decimalDomain = hexDomain.result
-        .map((element) => new BN(hexToDecimalString(element)))
-        .slice(1);
-
-      const stringDomain = useDecoded(decimalDomain);
-
-      return stringDomain;
-    } catch {
-      return Error('Could not get stark name');
-    }
-  }
-
-  public async getAddressFromStarkName(
-    name: string,
-    StarknetIdContract?: string
-  ): Promise<string | Error> {
-    const chainId = await this.getChainId();
-    const contract = StarknetIdContract ?? getStarknetIdContract(chainId);
-
-    try {
-      const addressData = await this.callContract({
-        contractAddress: contract,
-        entrypoint: 'domain_to_address',
-        calldata: compileCalldata({
-          domain: [useEncoded(name.replace('.stark', '')).toString(10)],
-        }),
-      });
-
-      return addressData.result[0];
-    } catch {
-      return Error('Could not get address from stark name');
-    }
   }
 
   public async estimateFee(
@@ -132,11 +88,9 @@ export class Account extends Provider implements AccountInterface {
       chainId,
     };
 
-    const signature = await this.signer.signTransaction(transactions, signerDetails);
-
-    const calldata = fromCallsToExecuteCalldata(transactions);
+    const invocation = await this.buildInvocation(transactions, signerDetails);
     const response = await super.getInvokeEstimateFee(
-      { contractAddress: this.address, calldata, signature },
+      { ...invocation },
       { version, nonce },
       blockIdentifier
     );
@@ -156,19 +110,14 @@ export class Account extends Provider implements AccountInterface {
     const nonce = toBN(providedNonce ?? (await this.getNonce()));
     const version = toBN(feeTransactionVersion);
     const chainId = await this.getChainId();
-    const contractDefinition = parseContract(contract);
 
-    const signature = await this.signer.signDeclareTransaction({
-      classHash,
-      senderAddress: this.address,
-      chainId,
-      maxFee: ZERO,
-      version,
-      nonce,
-    });
+    const payload = await this.buildDeclarePayload(
+      { classHash, contract },
+      { nonce, chainId, version, walletAddress: this.address, maxFee: ZERO }
+    );
 
     const response = await super.getDeclareEstimateFee(
-      { senderAddress: this.address, signature, contractDefinition },
+      { ...payload },
       { version, nonce },
       blockIdentifier
     );
@@ -192,23 +141,14 @@ export class Account extends Provider implements AccountInterface {
     const nonce = '0x0';
     const version = toBN(feeTransactionVersion);
     const chainId = await this.getChainId();
-    const contractAddress =
-      providedContractAddress ??
-      calculateContractAddressFromHash(addressSalt, classHash, constructorCalldata, 0);
 
-    const signature = await this.signer.signDeployAccountTransaction({
-      classHash,
-      contractAddress,
-      chainId,
-      maxFee: ZERO,
-      version,
-      nonce,
-      addressSalt,
-      constructorCalldata,
-    });
+    const payload = await this.buildAccountDeployPayload(
+      { classHash, addressSalt, constructorCalldata, contractAddress: providedContractAddress },
+      { nonce, chainId, version, walletAddress: this.address, maxFee: ZERO }
+    );
 
     const response = await super.getDeployAccountEstimateFee(
-      { classHash, addressSalt, constructorCalldata, signature },
+      { ...payload },
       { version, nonce },
       blockIdentifier
     );
@@ -224,29 +164,99 @@ export class Account extends Provider implements AccountInterface {
     payload: UniversalDeployerContractPayload | UniversalDeployerContractPayload[],
     transactionsDetail?: InvocationsDetails | undefined
   ): Promise<EstimateFee> {
-    const calls = [].concat(payload as []).map((it) => {
-      const {
-        classHash,
-        salt = '0',
-        unique = true,
-        constructorCalldata = [],
-      } = it as UniversalDeployerContractPayload;
-      const compiledConstructorCallData = compileCalldata(constructorCalldata);
+    const calls = this.buildUDCContractPayload(payload);
+    return this.estimateInvokeFee(calls, transactionsDetail);
+  }
 
+  public async estimateFeeBulk(
+    transactions: TransactionBulk,
+    { nonce: providedNonce, blockIdentifier }: EstimateFeeDetails = {}
+  ): Promise<EstimateFeeBulk> {
+    const nonce = toBN(providedNonce ?? (await this.getNonce()));
+    const version = toBN(feeTransactionVersion);
+    const chainId = await this.getChainId();
+
+    const params: any = await Promise.all(
+      [].concat(transactions as []).map(async (transaction: any, index: number) => {
+        const signerDetails: InvocationsSignerDetails = {
+          walletAddress: this.address,
+          nonce: toBN(Number(nonce) + index),
+          maxFee: ZERO,
+          version,
+          chainId,
+        };
+
+        const txPayload = transaction.payload;
+
+        let res;
+        if (typeof transaction === 'object' && transaction.type === 'INVOKE_FUNCTION') {
+          const invocation = await this.buildInvocation(
+            Array.isArray(txPayload) ? txPayload : [txPayload],
+            signerDetails
+          );
+          res = {
+            type: 'INVOKE_FUNCTION',
+            ...invocation,
+            version,
+            nonce: toBN(Number(nonce) + index),
+            blockIdentifier,
+          };
+        } else if (typeof transaction === 'object' && transaction.type === 'DECLARE') {
+          const declareContractPayload = await this.buildDeclarePayload(txPayload, signerDetails);
+          res = {
+            type: 'DECLARE',
+            ...declareContractPayload,
+            version,
+            nonce: toBN(Number(nonce) + index),
+            blockIdentifier,
+          };
+        } else if (typeof transaction === 'object' && transaction.type === 'DEPLOY_ACCOUNT') {
+          const payload = await this.buildAccountDeployPayload(txPayload, signerDetails);
+          res = {
+            type: 'DEPLOY_ACCOUNT',
+            ...payload,
+            version,
+            nonce,
+            blockIdentifier,
+          };
+        } else if (typeof transaction === 'object' && transaction.type === 'DEPLOY') {
+          const calls = this.buildUDCContractPayload(txPayload);
+          const invocation = await this.buildInvocation(calls, signerDetails);
+          res = {
+            type: 'INVOKE_FUNCTION',
+            ...invocation,
+            version,
+            nonce: toBN(Number(nonce) + index),
+            blockIdentifier,
+          };
+        }
+        return res;
+      })
+    );
+
+    const response = await super.getEstimateFeeBulk(params, blockIdentifier);
+
+    return [].concat(response as []).map((elem: any) => {
+      const suggestedMaxFee = estimatedFeeToMaxFee(elem.overall_fee);
       return {
-        contractAddress: UDC.ADDRESS,
-        entrypoint: UDC.ENTRYPOINT,
-        calldata: [
-          classHash,
-          salt,
-          toCairoBool(unique),
-          compiledConstructorCallData.length,
-          ...compiledConstructorCallData,
-        ],
+        ...elem,
+        suggestedMaxFee,
       };
     });
+  }
 
-    return this.estimateInvokeFee(calls, transactionsDetail);
+  public async buildInvocation(
+    call: Array<Call>,
+    signerDetails: InvocationsSignerDetails
+  ): Promise<Invocation> {
+    const calldata = fromCallsToExecuteCalldata(call);
+    const signature = await this.signer.signTransaction(call, signerDetails);
+
+    return {
+      contractAddress: this.address,
+      calldata,
+      signature,
+    };
   }
 
   public async execute(
@@ -497,5 +507,127 @@ export class Account extends Provider implements AccountInterface {
     }
 
     return feeEstimate.suggestedMaxFee.toString();
+  }
+
+  public async buildDeclarePayload(
+    { classHash, contract }: DeclareContractPayload,
+    { nonce, chainId, version, walletAddress, maxFee }: InvocationsSignerDetails
+  ): Promise<DeclareContractTransaction> {
+    const contractDefinition = parseContract(contract);
+    const signature = await this.signer.signDeclareTransaction({
+      classHash,
+      senderAddress: walletAddress,
+      chainId,
+      maxFee,
+      version,
+      nonce,
+    });
+
+    return {
+      senderAddress: walletAddress,
+      signature,
+      contractDefinition,
+    };
+  }
+
+  public async buildAccountDeployPayload(
+    {
+      classHash,
+      addressSalt = 0,
+      constructorCalldata = [],
+      contractAddress: providedContractAddress,
+    }: DeployAccountContractPayload,
+    { nonce, chainId, version, maxFee }: InvocationsSignerDetails
+  ): Promise<DeployAccountContractTransaction> {
+    const contractAddress =
+      providedContractAddress ??
+      calculateContractAddressFromHash(addressSalt, classHash, constructorCalldata, 0);
+
+    const signature = await this.signer.signDeployAccountTransaction({
+      classHash,
+      contractAddress,
+      chainId,
+      maxFee,
+      version,
+      nonce,
+      addressSalt,
+      constructorCalldata,
+    });
+
+    return {
+      classHash,
+      addressSalt,
+      constructorCalldata,
+      signature,
+    };
+  }
+
+  public buildUDCContractPayload(
+    payload: UniversalDeployerContractPayload | UniversalDeployerContractPayload[]
+  ): Call[] {
+    const calls = [].concat(payload as []).map((it) => {
+      const {
+        classHash,
+        salt = '0',
+        unique = true,
+        constructorCalldata = [],
+      } = it as UniversalDeployerContractPayload;
+      const compiledConstructorCallData = compileCalldata(constructorCalldata);
+
+      return {
+        contractAddress: UDC.ADDRESS,
+        entrypoint: UDC.ENTRYPOINT,
+        calldata: [
+          classHash,
+          salt,
+          toCairoBool(unique),
+          compiledConstructorCallData.length,
+          ...compiledConstructorCallData,
+        ],
+      };
+    });
+    return calls;
+  }
+
+  public async simulateTransaction(
+    calls: AllowArray<Call>,
+    { nonce: providedNonce, blockIdentifier }: EstimateFeeDetails = {}
+  ): Promise<TransactionSimulation> {
+    const transactions = Array.isArray(calls) ? calls : [calls];
+    const nonce = toBN(providedNonce ?? (await this.getNonce()));
+    const version = toBN(feeTransactionVersion);
+    const chainId = await this.getChainId();
+
+    const signerDetails: InvocationsSignerDetails = {
+      walletAddress: this.address,
+      nonce,
+      maxFee: ZERO,
+      version,
+      chainId,
+    };
+
+    const invocation = await this.buildInvocation(transactions, signerDetails);
+    const response = await super.getSimulateTransaction(
+      invocation,
+      { version, nonce },
+      blockIdentifier
+    );
+
+    const suggestedMaxFee = estimatedFeeToMaxFee(response.fee_estimation.overall_fee);
+
+    return {
+      ...response,
+      fee_estimation: {
+        ...response.fee_estimation,
+        suggestedMaxFee,
+      },
+    };
+  }
+
+  public override async getStarkName(
+    address: BigNumberish = this.address, // default to the wallet address
+    StarknetIdContract?: string
+  ): Promise<string> {
+    return super.getStarkName(address, StarknetIdContract);
   }
 }