+++ www/docusaurus.config.js
@@ -7,7 +7,7 @@ const darkCodeTheme = require('prism-react-renderer/themes/dracula');
 /** @type {import('@docusaurus/types').Config} */
 const config = {
   title: 'Starknet.js',
-  tagline: 'JavaScript libray for StarkNet',
+  tagline: 'JavaScript library for Starknet',
   url: 'https://starknetjs.com',
   baseUrl: '/',
   onBrokenLinks: 'throw',
@@ -16,17 +16,17 @@ const config = {
   organizationName: '0xs34n', // Usually your GitHub org/user name.
   projectName: 'starknet.js', // Usually your repo name.
 
-  plugins: [
-    [
-      '@docusaurus/plugin-content-docs',
-      {
-        id: 'guides',
-        path: 'guides',
-        routeBasePath: 'guides',
-        sidebarPath: require.resolve('./sidebars.js'),
-      },
-    ],
-  ],
+  // plugins: [
+  //   [
+  //     '@docusaurus/plugin-content-docs',
+  //     {
+  //       id: 'guides',
+  //       path: 'guides',
+  //       routeBasePath: 'guides',
+  //       sidebarPath: require.resolve('./sidebars.js'),
+  //     },
+  //   ],
+  // ],
 
   presets: [
     [
@@ -69,7 +69,14 @@ const config = {
           {
             position: 'left',
             label: 'Guides',
-            to: '/guides/intro',
+            type: 'doc',
+            docId: 'guides/intro',
+          },
+          {
+            type: 'docsVersionDropdown',
+            position: 'left',
+            // dropdownItemsAfter: [{ to: '/versions', label: 'All versions' }],
+            dropdownActiveClassDisabled: true,
           },
           // {to: '/blog', label: 'Blog', position: 'left'},
           {
@@ -91,7 +98,7 @@ const config = {
               },
               {
                 label: 'Guides',
-                to: '/guides/intro',
+                to: '/docs/guides/intro',
               },
             ],
           },